enable some warnings on windows and gcc, fixes for them

+++ tunnels/TcpConnector/upstream/init.c
@@ -72,7 +72,7 @@ void tcpconnectorTunnelUpStreamInit(tunnel_t *t, line_t *l)
     // sockaddr_set_ipport(&(dest_ctx.addr), "127.0.0.1", 443);
 
     wloop_t *loop   = getWorkerLoop(getWID());
-    int      sockfd = socket(dest_ctx->ip_address.type, SOCK_STREAM, 0);
+    int      sockfd = (int) socket(dest_ctx->ip_address.type, SOCK_STREAM, 0);
 
     if (sockfd < 0)
     {
